+++ pkg/shelf/CHANGELOG.md
@@ -3,6 +3,8 @@
 * Add a `Cascade` helper that runs handlers in sequence until one returns a
   response that's neither a 404 nor a 405.
 
+* Add a `Request.change` method that copies a request with new header values.
+
 * Add a `Request.hijack` method that allows handlers to gain access to the
   underlying HTTP socket.